GH60 simplification of VacancyScreen.kt
