+++ app/src/main/java/ru/practicum/android/diploma/vacancy/ui/components/VacancyStates.kt
@@ -1,0 +1,57 @@
+package ru.practicum.android.diploma.vacancy.ui.components
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import ru.practicum.android.diploma.R
+import ru.practicum.android.diploma.core.data.dto.VacancyDetail
+import ru.practicum.android.diploma.vacancy.ui.viewmodel.VacancyViewModel
+
+@Composable
+fun ServerErrorState() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Image(
+            painter = painterResource(R.drawable.ph_server_error_vacancy_screen),
+            contentDescription = "Server error",
+        )
+    }
+}
+
+@Composable
+fun ErrorState() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Image(
+            painter = painterResource(R.drawable.ph_error_vacancy_screen),
+            contentDescription = "Server error",
+        )
+    }
+}
+
+@Composable
+fun SuccessState(vacancy: VacancyDetail, viewModel: VacancyViewModel) {
+    VacancyHeader(vacancy)
+    EmployerCard(vacancy)
+    ExperienceSection(vacancy)
+    DescriptionSection(vacancy, viewModel)
+}
+
+@Composable
+fun LoadingState() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        CircularProgressIndicator()
+    }
+}
